Added distinction between "database" and server state
